Improve App Group selection UI.
Now, in app list, there's a separate bottom toolbar shown, which contains
current app group name and prev/next buttons. Clicking on group name
will show list of groups to switch to (as an optimization of screen
space, this list lacks "New group..." choice).

A bit of regression is that last selected group is not saved in app
preferences (and thus next time launcher starts, it won't be preselected).
This is considered a minor regression (can be consider speed/storage usage
optimizations ;-) ).